finish JVMMetrics

+++ framework/src/main/java/edu/uci/eecs/wukong/framework/api/metrics/Counter.java
@@ -19,7 +19,7 @@ public class Counter implements Metrics {
 		return inc(1);
 	}
 	
-	public long inc(int n) {
+	public long inc(long n) {
 		return count.addAndGet(n);
 	}
